move `EntityBorrowMut` into `entity_ref.rs`
